Imporved "showHint" state functionality.
This allows keypresses to be handled normally in
the "showHint" state in the same way that they
would be handled in the "showQuestion" state.
This includes things like pressing the enter key
or space bar to show the answer.

This patch also shifts from defining the answer
key as a string to as a Qt Key Enum, making it
simpler to use non-printable keys.